+++ src/features/mandatory/removeGloballyBannedItems/constants.ts
@@ -4,15 +4,15 @@ export const BANNED_COLLECTIBLES = new Set<CollectibleType>([
 ]);
 
 export const BANNED_COLLECTIBLES_ON_SEEDED_RACES = new Set<CollectibleType>([
-  // Sol is mostly useless if you start with the Compass
-  CollectibleType.COLLECTIBLE_SOL,
-
   // Since drops are seeded and given in order, Glyph of Balance does not work properly
-  CollectibleType.COLLECTIBLE_GLYPH_OF_BALANCE,
+  CollectibleType.COLLECTIBLE_GLYPH_OF_BALANCE, // 464
 
   // Damocles is unseeded
-  CollectibleType.COLLECTIBLE_DAMOCLES,
-  CollectibleType.COLLECTIBLE_DAMOCLES_PASSIVE,
+  CollectibleType.COLLECTIBLE_DAMOCLES, // 577
+  CollectibleType.COLLECTIBLE_DAMOCLES_PASSIVE, // 656
+
+  // Sol is mostly useless if you start with the Compass
+  CollectibleType.COLLECTIBLE_SOL, // 588
 ]);
 
 export const BANNED_COLLECTIBLES_WITH_VOID = new Set<CollectibleType>([