+++ src/features/optional/gameplay/extraStartingItems/replacePlaceholdersOnEden.ts
@@ -1,13 +1,12 @@
 import { getPlayersOfType, getRandomArrayElement } from "isaacscript-common";
 import g from "../../../../globals";
 import { PASSIVE_ITEMS_FOR_EDEN } from "../../../../passiveItemsForEden";
+import { CollectibleTypeCustom } from "../../../../types/CollectibleTypeCustom";
 import { giveCollectibleAndRemoveFromPools } from "../../../../utilGlobals";
 import * as showEdenStartingItems from "../../characters/showEdenStartingItems";
 import { COLLECTIBLE_REPLACEMENT_MAP } from "./constants";
 
 export function postGameStarted(): void {
-  const startSeed = g.seeds.GetStartSeed();
-
   const edens = getPlayersOfType(
     PlayerType.PLAYER_EDEN,
     PlayerType.PLAYER_EDEN_B,
@@ -19,12 +18,28 @@ export function postGameStarted(): void {
       }
 
       player.RemoveCollectible(placeholderItem);
-      const newCollectible = getRandomArrayElement(
-        PASSIVE_ITEMS_FOR_EDEN,
-        startSeed,
+      const replacementCollectibleType =
+        getEdenReplacementCollectibleType(player);
+      showEdenStartingItems.changeStartingPassiveItem(
+        replacementCollectibleType,
       );
-      showEdenStartingItems.changeStartingPassiveItem(newCollectible);
-      giveCollectibleAndRemoveFromPools(player, newCollectible);
+      giveCollectibleAndRemoveFromPools(player, replacementCollectibleType);
     }
   }
+}
+
+export function getEdenReplacementCollectibleType(
+  player: EntityPlayer,
+): CollectibleType | CollectibleTypeCustom {
+  const startSeed = g.seeds.GetStartSeed();
+
+  let replacementCollectible: CollectibleType | CollectibleTypeCustom;
+  do {
+    replacementCollectible = getRandomArrayElement(
+      PASSIVE_ITEMS_FOR_EDEN,
+      startSeed,
+    );
+  } while (!player.HasCollectible(replacementCollectible));
+
+  return replacementCollectible;
 }